Remove health file at old location.
