+++ utilities/src/with_std/health.rs
@@ -10,8 +10,6 @@ use serde::Deserialize;
 use tracing::info;
 use warp::Filter;
 
-// use crate::settings;
-
 #[derive(Debug, Deserialize, Clone, Default, PartialEq, Eq)]
 pub struct HealthCheck {
 	pub hostname: String,
@@ -52,56 +50,54 @@ pub async fn start<'a, 'env>(
 	Ok(())
 }
 
-// #[cfg(test)]
-// mod tests {
+#[cfg(test)]
+mod tests {
 
-// 	use futures_util::FutureExt;
+	use futures_util::FutureExt;
 
-// 	use crate::settings::Settings;
+	use super::*;
 
-// 	use super::*;
+	#[tokio::test]
+	async fn health_check_test() {
+		let health_check = HealthCheck { hostname: "127.0.0.1".to_string(), port: 5555 };
 
-// 	#[tokio::test]
-// 	async fn health_check_test() {
-// 		let health_check = Settings::new_test().unwrap().health_check.unwrap();
+		task_scope::task_scope(|scope| {
+			async {
+				let has_completed_initialising =
+					Arc::new(std::sync::atomic::AtomicBool::new(false));
+				start(scope, &health_check, has_completed_initialising.clone()).await.unwrap();
 
-// 		task_scope::task_scope(|scope| {
-// 			async {
-// 				let has_completed_initialising =
-// 					Arc::new(std::sync::atomic::AtomicBool::new(false));
-// 				start(scope, &health_check, has_completed_initialising.clone()).await.unwrap();
+				let request_test = |path: &'static str,
+				                    expected_status: reqwest::StatusCode,
+				                    expected_text: &'static str| {
+					let health_check = health_check.clone();
 
-// 				let request_test = |path: &'static str,
-// 				                    expected_status: reqwest::StatusCode,
-// 				                    expected_text: &'static str| {
-// 					let health_check = health_check.clone();
+					async move {
+						let resp = reqwest::get(&format!(
+							"http://{}:{}/{}",
+							&health_check.hostname, &health_check.port, path
+						))
+						.await
+						.unwrap();
 
-// 					async move {
-// 						let resp = reqwest::get(&format!(
-// 							"http://{}:{}/{}",
-// 							&health_check.hostname, &health_check.port, path
-// 						))
-// 						.await
-// 						.unwrap();
+						assert_eq!(expected_status, resp.status());
+						assert_eq!(resp.text().await.unwrap(), expected_text);
+					}
+				};
 
-// 						assert_eq!(expected_status, resp.status());
-// 						assert_eq!(resp.text().await.unwrap(), expected_text);
-// 					}
-// 				};
+				// starts with `has_completed_initialising` set to false
+				request_test("health", reqwest::StatusCode::OK, INITIALISING).await;
+				request_test("invalid", reqwest::StatusCode::NOT_FOUND, "").await;
 
-// 				// starts with `has_completed_initialising` set to false
-// 				request_test("health", reqwest::StatusCode::OK, INITIALISING).await;
-// 				request_test("invalid", reqwest::StatusCode::NOT_FOUND, "").await;
+				has_completed_initialising.store(true, std::sync::atomic::Ordering::Relaxed);
 
-// 				has_completed_initialising.store(true, std::sync::atomic::Ordering::Relaxed);
+				request_test("health", reqwest::StatusCode::OK, RUNNING).await;
 
-// 				request_test("health", reqwest::StatusCode::OK, RUNNING).await;
-
-// 				Ok(())
-// 			}
-// 			.boxed()
-// 		})
-// 		.await
-// 		.unwrap();
-// 	}
-// }
+				Ok(())
+			}
+			.boxed()
+		})
+		.await
+		.unwrap();
+	}
+}
